+++ LinkDotNet.BuildInformation.Sample/Program.cs
@@ -8,3 +8,4 @@ Console.WriteLine($"Assembly version: {BuildInformation.AssemblyVersion}");
 Console.WriteLine($"Assembly file version: {BuildInformation.AssemblyFileVersion}");
 Console.WriteLine($"Assembly name: {BuildInformation.AssemblyName}");
 Console.WriteLine($"Target framework moniker: {BuildInformation.TargetFrameworkMoniker}");
+Console.WriteLine($"Analysis level: {BuildInformation.Nullability}");